+++ app/components/eto/overview/EtoOverviewStatus/ClaimWidget.module.scss
@@ -1,0 +1,12 @@
+.claim-widget {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.message {
+  max-width: 300px;
+  text-align: center;
+  margin-bottom: px-to-rem(20px);
+}
