+++ app/components/eto/overview/EtoOverviewStatus/CounterWidget.module.scss
@@ -1,0 +1,19 @@
+.counter-widget {
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+}
+
+.title {
+  @include font-default;
+
+  margin-bottom: px-to-rem(8px);
+  text-transform: uppercase;
+}
+
+.zone {
+  @include font-label;
+
+  color: $color-orange;
+  margin-bottom: px-to-rem(8px);
+}
