+++ app/config/yupLocales.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 import { FormattedMessage } from "react-intl-phraseapp";
-import { setLocale } from "yup";
 
 const mixed = {
   default: (values: any) => <FormattedMessage id={"form.field.error.default"} values={values} />,
