+++ app/styles/elements/a.scss
@@ -1,6 +1,5 @@
 a {
   &[href^="http"],
-  &[href^="/"],
   &[download] {
     text-decoration: underline;
   }
